requirements.txt
discord.py
aiosqlite
nest_asyncio
beautifulsoup4
requests